dynamic components moved from ssr to user client
