+++ src/app/(public)/search/page.tsx
@@ -1,17 +1,11 @@
 import type { Metadata } from 'next'
-import dynamic from 'next/dynamic'
-import { SkeletonLoader } from '@/ui/SkeletonLoader'
 import { NO_INDEX_PAGE } from '@/constants/seo.constants'
-
-const DynamicSearchPage = dynamic(
-	() => import('./SearchClientPage').then(mod => mod.SearchClientPage),
-	{ ssr: false, loading: () => <SkeletonLoader className='w-10 mb-0 rounded-md' /> }
-)
+import { SearchSubPage } from './SearchSubPage'
 
 export const metadata: Metadata = {
 	title: 'Search',
 	...NO_INDEX_PAGE
 }
 export default function SearchPage() {
-	return <DynamicSearchPage />
+	return <SearchSubPage />
 }
