+++ src/app/(user)/subscriptions/page.tsx
@@ -1,16 +1,6 @@
 import type { Metadata } from 'next'
-import dynamic from 'next/dynamic'
-import { SkeletonLoader } from '@/ui/SkeletonLoader'
 import { NO_INDEX_PAGE } from '@/constants/seo.constants'
-
-const DynamicSubscriptionsPage = dynamic(
-	() => import('./SubscriptionsPage').then(mod => mod.SubscriptionsPage),
-	{
-		//[FIXME]?
-		// ssr: false,
-		loading: () => <SkeletonLoader className='w-10 mb-0 rounded-md' />
-	}
-)
+import { SubscriptionsSubPage } from './SubscriptionsSubPage'
 
 export const metadata: Metadata = {
 	title: 'Subscriptions',
@@ -18,5 +8,5 @@ export const metadata: Metadata = {
 }
 
 export default function Subscriptions() {
-	return <DynamicSubscriptionsPage />
+	return <SubscriptionsSubPage />
 }
